Adding service method to get the payload from external API

+++ Adp.Eai.Domain/Models/Calculation.cs
@@ -17,6 +17,6 @@
         /// <summary>
         /// 
         /// </summary>
-        public long Result { get; set; }
+        public decimal Result { get; set; }
     }
 }
